// studio-web/src/api/models/emailTemplateDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: v1
 */

export interface EmailTemplateDTO {
  /** @nullable */
  id?: string | null;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  htmlContent: string;
  /** @nullable */
  textContent?: string | null;
  createdAt?: string;
  updatedAt?: string;
  /** @nullable */
  createdBy?: string | null;
}